+++ .appbuilder-state.json
@@ -1,1 +1,1 @@
-{"applicationId":"5qieqhq7ow24","applicationName":"Bugs after Ignite UI 15 update","platform":"Blazor"}
+{"applicationId":"5qieqhq7ow24","applicationName":"🐞 Bugs after Ignite UI 15 update","platform":"Blazor"}
