+++ Bugs after Ignite UI 15 update/Models/MovieAppData1MovieList.cs
@@ -1,0 +1,7 @@
+//namespace Bugs_after_Ignite_UI_15_update.Data.MovieAppData1; // Razor won't recognize third level namespace
+namespace Bugs_after_Ignite_UI_15_update.MovieAppData1;
+
+public class MovieListType
+{
+    public string Name { get; set; }
+}
