+++ Bugs after Ignite UI 15 update/Models/MovieAppData1NowPlaying.cs
@@ -1,0 +1,10 @@
+//namespace Bugs_after_Ignite_UI_15_update.Data.MovieAppData1; // Razor won't recognize third level namespace
+namespace Bugs_after_Ignite_UI_15_update.MovieAppData1;
+
+public class NowPlayingType
+{
+    public string Name { get; set; }
+    public string Genre { get; set; }
+    public string Description { get; set; }
+    public string MoviePoster { get; set; }
+}
