+++ Bugs after Ignite UI 15 update/Models/MovieAppData1TheatresNearYou.cs
@@ -1,5 +1,5 @@
-//namespace Bugs_after_Ignite_UI_15_update.Data.MovieAppData; // Razor won't recognize third level namespace
-namespace Bugs_after_Ignite_UI_15_update.MovieAppData;
+//namespace Bugs_after_Ignite_UI_15_update.Data.MovieAppData1; // Razor won't recognize third level namespace
+namespace Bugs_after_Ignite_UI_15_update.MovieAppData1;
 
 public class TheatresNearYouType
 {
