--- Bugs after Ignite UI 15 update/Models/MovieAppDataTheatres.cs
@@ -1,7 +1,0 @@
-//namespace Bugs_after_Ignite_UI_15_update.Data.MovieAppData; // Razor won't recognize third level namespace
-namespace Bugs_after_Ignite_UI_15_update.MovieAppData;
-
-public class TheatresType
-{
-    public string Theatres { get; set; }
-}
