+++ Bugs after Ignite UI 15 update/Pages/ImgResizing.razor
@@ -1,4 +1,7 @@
 @page "/img-resizing"
+@using MovieAppData
+@using MovieAppData1
+@using Financial
 
 <div class="row-layout img-resizing-container">
     <div class="column-layout group">
