+++ Bugs after Ignite UI 15 update/Program.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 using Bugs_after_Ignite_UI_15_update;
 using Bugs_after_Ignite_UI_15_update.MovieAppData;
+using Bugs_after_Ignite_UI_15_update.MovieAppData1;
 using Bugs_after_Ignite_UI_15_update.Financial;
 using IgniteUI.Blazor.Controls;
 
@@ -11,6 +12,7 @@ builder.RootComponents.Add<HeadOutlet>("head::after");
 
 builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
 builder.Services.AddScoped<IMovieAppDataService>(sp => new MovieAppDataService(new HttpClient {BaseAddress = new Uri(builder.HostEnvironment.BaseAddress)}));
+builder.Services.AddScoped<IMovieAppData1Service>(sp => new MovieAppData1Service(new HttpClient {BaseAddress = new Uri(builder.HostEnvironment.BaseAddress)}));
 builder.Services.AddScoped<IFinancialService>(sp => new FinancialService(new HttpClient {BaseAddress = new Uri(builder.HostEnvironment.BaseAddress)}));
 RegisterIgniteUI(builder.Services);
 
@@ -23,9 +25,9 @@ void RegisterIgniteUI(IServiceCollection services)
         typeof(IgbIconButtonModule),
         typeof(IgbRippleModule),
         typeof(IgbNavDrawerModule),
-        typeof(IgbButtonModule),
         typeof(IgbTabsModule),
         typeof(IgbCardModule),
+        typeof(IgbButtonModule),
         typeof(IgbDatePickerModule),
         typeof(IgbListModule),
         typeof(IgbAvatarModule)
