+++ Bugs after Ignite UI 15 update/Services/IMovieAppData1Service.cs
@@ -1,0 +1,15 @@
+namespace Bugs_after_Ignite_UI_15_update.MovieAppData1
+{
+    public interface IMovieAppData1Service
+    {
+        Task<List<NowPlayingType>> GetNowPlaying();
+
+        Task<List<MovieListType>> GetMovieList();
+
+        Task<List<TheatresType>> GetTheatres();
+
+        Task<List<ShowtimesType>> GetShowtimes();
+
+        Task<List<TheatresNearYouType>> GetTheatresNearYou();
+    }
+}
