+++ Bugs after Ignite UI 15 update/wwwroot/index.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
-    <title>Bugs after Ignite UI 15 update</title>
+    <title>🐞 Bugs after Ignite UI 15 update</title>
     <base href="/" />
     <!-- <link href="css/bootstrap/bootstrap.min.css" rel="stylesheet" /> -->
     <link href="_content/IgniteUI.Blazor/themes/light/material.css" rel="stylesheet" />
