+++ TestBugs after Ignite UI 15 update/Pages/TestMovieAppRebuilt.cs
@@ -1,11 +1,11 @@
 using Bunit;
 using Microsoft.Extensions.DependencyInjection;
 using Bugs_after_Ignite_UI_15_update.Pages;
-using Bugs_after_Ignite_UI_15_update.MovieAppData1;
+using Bugs_after_Ignite_UI_15_update.MovieAppData;
 
 namespace TestBugs_after_Ignite_UI_15_update
 {
-	public class TestMovieApp
+	public class TestMovieAppRebuilt
 	{
 		[Fact]
 		public void ViewIsCreated()
@@ -13,16 +13,16 @@ namespace TestBugs_after_Ignite_UI_15_update
 			using var ctx = new TestContext();
 			ctx.JSInterop.Mode = JSRuntimeMode.Loose;
 			ctx.Services.AddIgniteUIBlazor(
-				typeof(IgbButtonModule),
-				typeof(IgbRippleModule),
 				typeof(IgbTabsModule),
 				typeof(IgbCardModule),
+				typeof(IgbButtonModule),
+				typeof(IgbRippleModule),
 				typeof(IgbIconButtonModule),
 				typeof(IgbDatePickerModule),
 				typeof(IgbListModule),
 				typeof(IgbAvatarModule));
-			ctx.Services.AddScoped<IMovieAppData1Service>(sp => new MockMovieAppData1Service());
-			var componentUnderTest = ctx.RenderComponent<MovieApp>();
+			ctx.Services.AddScoped<IMovieAppDataService>(sp => new MockMovieAppDataService());
+			var componentUnderTest = ctx.RenderComponent<MovieAppRebuilt>();
 			Assert.NotNull(componentUnderTest);
 		}
 	}
